git-svn-id: https://code.crt.realtors.org/svn/odbcretsdrv/ezrets/branches/build-64-bit@1399 fe7ac059-47f5-0310-8599-d0339b4e125c

+++ project/installer/src/RegisterFuncs.cpp
@@ -37,7 +37,7 @@ StrVector drivers2vector(LPCSTR drivers)
     
     while (*i)
     {
-        int len = strlen(i);
+        size_t len = strlen(i);
         string drv(i, len);
 
         drvvector.push_back(drv);
